requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2024.1
dj-database-url==2.2.0
Django==5.0.4
django-address==0.2.8
django-admin-material-dashboard==1.0.10
django-autoslug==1.9.9
django-braces==1.15.0
django-cities==0.6.2
django-cities-light==3.10.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-countries==7.6.1
django-crispy-forms==2.1
django-crontab==0.7.1
django-embed-video==1.4.4
django-filter==24.2
django-form-surveys==2.2.2
django-htmx==1.18.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-likert-field==0.2.0
django-localflavor==4.0
django-mail-admin==0.3.2
django-model-utils==4.5.1
django-phone-field==1.8.1
django-phonenumber-field==7.3.0
django-product-review==0.0.2
django-social-share==2.3.0
django-star-ratings==0.9.2
django-taggit==5.0.1
django-theme==0.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
flower==1.1.0
gunicorn==22.0.0
h11==0.14.0
humanize==4.10.0
idna==3.7
joblib==1.4.2
jsonfield==3.1.0
kombu==5.3.7
nltk==3.8.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pika==1.3.2
pillow==10.3.0
progressbar2==4.4.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-stdnum==1.20
python-utils==3.8.2
pytz==2024.1
redis==5.0.7
regex==2024.5.15
requests==2.31.0
setuptools==69.5.1
simplejwt==2.0.1
six==1.16.0
sqlparse==0.5.0
swapper==1.3.0
sweetify==2.3.1
textblob==0.18.0.post0
tornado==6.4.1
tqdm==4.66.4
typing==3.7.4.3
typing_extensions==4.12.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.30.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
